# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Velora DNS Management API",
    "version": "1.0.0",
    "description": "Session cookies require X-CSRF-Token on mutations. API bearer tokens bypass CSRF and require the documented scope. All API responses use a data or error envelope except text/dns zone export. Secrets are never returned by read endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "paths": {
    "/api/v1/analytics": {
      "get": {
        "operationId": "get_analytics",
        "summary": "Get analytics",
        "tags": [
          "analytics"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Analytics"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "viewer-or-higher"
      }
    },
...
    "/api/v1/queries": {
      "get": {
        "operationId": "get_queries",
        "summary": "Get queries",
        "tags": [
          "queries"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Query"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "viewer-or-higher"
      }
    },
    "/api/v1/query-stats": {
      "get": {
        "operationId": "get_query_stats",
        "summary": "Get query-stats",
        "tags": [
          "query-stats"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/QuerySummary"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "viewer-or-higher"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "velora_ API token; scopes read, write, or admin"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "velora_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required for cookie-authenticated mutations"
      }
    },
    "schemas": {
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Configuration path such as dns.allowed_clients[1]"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Invalid or restart-only configuration fields (invalid_config, config_requires_restart)"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Query": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time"
          },
          "client_ip": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "rcode": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "source": {
            "type": "string"
          },
          "upstream": {
            "type": "string"
          },
          "cache_hit": {
            "type": "boolean"
          },
          "client_name": {
            "type": "string",
            "description": "Friendly name of the matching client definition, when any"
          }
        },
        "required": [
          "id",
          "domain",
          "type",
          "rcode",
          "source"
        ]
      },
      "QueryRanking": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "Client name for top_clients entries, when known"
          }
        },
        "required": [
          "value",
          "count"
        ]
      },
      "AnalyticsBucket": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "blocked": {
            "type": "integer",
            "format": "int64"
          },
          "cached": {
            "type": "integer",
            "format": "int64"
          },
          "failed": {
            "type": "integer",
            "description": "SERVFAIL answers"
          },
          "average_ms": {
            "type": "number",
            "description": "Average upstream response time of answered queries"
          }
        },
        "required": [
          "start",
          "total",
          "blocked",
          "cached",
          "failed",
          "average_ms"
        ]
      },
      "UpstreamUsage": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "queries": {
            "type": "integer",
            "format": "int64"
          },
          "failed": {
            "type": "integer",
            "format": "int64"
          },
          "average_ms": {
            "type": "number"
          }
        },
        "required": [
          "address",
          "queries",
          "failed",
          "average_ms"
        ]
      },
      "Analytics": {
        "type": "object",
        "properties": {
          "range": {
            "type": "string",
            "enum": [
              "1h",
              "24h",
              "7d",
              "30d"
            ]
          },
          "window_start": {
            "type": "string",
            "format": "date-time"
          },
          "window_end": {
            "type": "string",
            "format": "date-time"
          },
          "bucket_seconds": {
            "type": "integer",
            "format": "int64"
          },
          "history_start": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Oldest retained query"
          },
          "totals": {
            "$ref": "#/components/schemas/AnalyticsBucket"
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalyticsBucket"
            }
          },
          "query_types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryRanking"
            }
          },
          "response_codes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryRanking"
            }
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryRanking"
            }
          },
          "upstreams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpstreamUsage"
            }
          },
          "top_blocked": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryRanking"
            }
          }
        },
        "required": [
          "range",
          "window_start",
          "window_end",
          "bucket_seconds",
          "history_start",
          "totals",
          "series",
          "query_types",
          "response_codes",
          "sources",
          "upstreams",
          "top_blocked"
        ]
      },
      "QuerySummary": {
        "type": "object",
        "properties": {
          "window_start": {
            "type": "string",
            "format": "date-time"
          },
          "window_end": {
            "type": "string",
            "format": "date-time"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "blocked": {
            "type": "integer",
            "format": "int64"
          },
          "top_domains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryRanking"
            }
          },
          "top_clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryRanking"
            }
          }
        }
      },
...
    }
  }
}